The height of contentContainer's child should be determined by the height of its own children
Previously the height of contentContainer's child was set to match contentContainer's height. This is not desirable. contentContainer is intended to have a scrollbar so you can browse around its tall child.

Now the height of contentContainer's child is determined by the height of its own children. Consequently, it can expand as much as it wants and the user can use contentContainer's scrollbar to browse through its child.

The only effect I noticed of this bug was that the last line of podcast descriptions and chapters & transcript was against the edge of the browser window. Now there's some room to render the last line's bottom margin.